animation-helper/Animation/requirements.txt
opencv-python
rembg
pillow
numpy
streamlit
imageio
streamlit-drawable-canvas
